add options to binary_to_term
binary_to_term is a powerful tool for easily shuttling data over the network.
It is not suitable for use with untrusted data, as it happily leaks atoms and
transfers data types that can leak atoms as well (funs, pids, and refs include
an atom internally for the node name).

After a short discussion on the list, consensus appears to be that
binary_to_term needed options.  This patch implements the second form of
binary_to_term that takes options, and a simple option called 'safe'.  If
specified, this option will refuse to decode any dangerous data types and will
not leak atoms.

Note that it will decode funs, pids, and refs safely, so it should be able to
encode / decode these fine unless they are carefully crafted to leak atoms or
possibly if they are persisted between invocations of the VM.  As a special
case, decoding done during distribution doesn't have these limitations either.

In short, it should "just work" most of the time.